requirements.txt
openai
anthropic 
mistralai 
gTTS 
streamlit
python-dotenv